Battle

Description

This module deals with attacking and defending types and gives an interface to calculate battle outcomes.

Documentation

data Defenders

Constructors

OneDef

TwoDef

data Attackers

Constructors

OneAtt

TwoAtt

ThreeAtt

doBattle :: Attackers -> Defenders -> StdGen -> (Int, Int, StdGen)

Takes a number of attackers (Attackers) and a number of defenders (Defenders) and calculates the attackers lost, the defeners lost and the new StdGen (returned in a tuple in that order).