GameElements

Description

This module holds the types that represent individual game pieces and other game-specific enumerated types.

Documentation

data Player

Each player is represented by the colour they play as.

Constructors

Black

Blue

Green

Red

Yellow

data Card

We ignore the country shown on the card.

Constructors

Infantry

Cavalry

Artillery

Wild

type CardSet = (Card, Card, Card)

Cards always come in sets of 3 when being traded in.

data TradeIn

A player csan trade in at most 2 sets at once.

Constructors

None

OneSet CardSet

TwoSet CardSet CardSet