Interface

Description

This module wraps many of the functions in moves in additional validation logic and exposes a Hasskell version of the JSON interface used to communicate with the Server.

Documentation

data Game

The main type holding all the needed information about the game.

data Response

Constructors

General Update

Special Question Player

Invalid Error Player

GameWon Player

data Request

Constructors

Request Player RequestType

empty :: StdGen -> Game

Creates a new empty game with a given StdGen.

addPlayer :: Game -> (Player, Game)

Errors if we aren't in the waiting phase, or if the waiting room is full, or if the given waiting room is invalid (i.e. there are repeats)

receive :: Request -> Game -> (Response, Game)

Total function. For any Request, checks whether it is valid and returns an appropriate Response message along with the new Game with the correct changes made.