Moves

Description

The basic implementation of the rules for the game, verifying individual actions that manipulate the state.

Documentation

data Country

The Country type should be inhabited by 42 nullary constructors with the names corresponding to the country names (including capitalisations, missing spaces where necessary).

Constructors

Alaska

Alberta

CentralAmerica

EasternUnitedStates

Greenland

NorthwestTerritory

Ontario

Quebec

WesternUnitedStates

Argentina

Brazil

Peru

Venezuela

GreatBritain

Iceland

NorthernEurope

Scandinavia

SouthernEurope

Ukraine

WesternEurope

Congo

EastAfrica

Egypt

Madagascar

NorthAfrica

SouthAfrica

Afghanistan

China

India

Irkutsk

Japan

Kamchatka

MiddleEast

Mongolia

Siam

Siberia

Ural

Yakutsk

EasternAustralia

Indonesia

NewGuinea

WesternAustralia

reinforce :: TradeIn -> [(Country, Int)] -> GameState -> Maybe GameState

Add reinforcements (specified as a list of pairs of Country and non-negative Int number of troops to add). Must be during the correct phase. Should update phase to Attack Normal. Continent and country bonuses should be taken into account.

attack :: Country -> Country -> Attackers -> GameState -> Maybe GameState

Puts the state into the MidBattle MiniPhase, which records the attacking and defending country, and the number of attackers. Must Be neighbouring countries owned by different players. Must be during the correct phase. Attacker must have < number of attackers in the attacking country.

chooseDefenders :: Defenders -> GameState -> Maybe GameState

Only defined for MidBattle MiniPhase. Chooses the number of defenders the defending country uses. The defending country must have >= number of defenders used. Enters WonBattle phase if there are no defenders left, otherwise goes back to Attack Normal MiniPhase.

invade :: Int -> GameState -> Maybe GameState

Invades a country from another with Int number of troops, meaning owner is changed and Int troops are tranferred from the attacking country to the defending country. Should be only able to be called in WonBattle MiniPhase. Can't invade with less than the number of attackers left. Should move back to a Normal Attack MiniPhase. Gives an error if called with a MiniPhase which shouldn't be possible. Kicks players out if their last country is invaded. Also calculates when winning. It must invade with at least the number of attackers left, must leave at least 1 in the previous country.

endAttack :: GameState -> Maybe GameState

Must be during the Attack Normal MiniPhase. Should update phase to Fortify.

fortify :: Country -> Country -> Int -> GameState -> Maybe GameState

Moves troops from one country to another. Countries must be neighbours and owned by the current player. Must be during the correct phase. Troops are sent from the first country to the second one. Should update phase to Reinforce and call nextTurn.

skipFortify :: GameState -> Maybe GameState

Must be during the correct phase. Should update phase to Reinforce and call nextTurn.

trade :: TradeIn -> [(Country, Int)] -> GameState -> Maybe GameState

Only valid during the Attack TimeToTrade MiniPhase. Takes a TradeIn (owned by the current player) and caches it in for troops.