Parse

Description

This module handles the encoding and decoding of JSON responses and requests.

Documentation

decodeRequest :: ByteString -> Either Request ParseError

Reads in a received ByteString and returns a Left Request or a Right ParseError.

encodeResponse :: Response -> ByteString

Encodes a Response to a ByteString.

type ParseError = ByteString

Orphan instances

ToJSON Response

Instance details

Methods

toJSON :: Response -> Value

toEncoding :: Response -> Encoding

toJSONList :: [Response] -> Value

toEncodingList :: [Response] -> Encoding

FromJSON Request

Instance details

Methods

parseJSON :: Value -> Parser Request

parseJSONList :: Value -> Parser [Request]