RiskBoard

Description

This module deals with representing the game board itself and also provides the Country type.

Documentation

data Country

The Country type should be inhabited by 42 nullary constructors with the names corresponding to the country names (including capitalisations, missing spaces where necessary).

Constructors

Alaska

Alberta

CentralAmerica

EasternUnitedStates

Greenland

NorthwestTerritory

Ontario

Quebec

WesternUnitedStates

Argentina

Brazil

Peru

Venezuela

GreatBritain

Iceland

NorthernEurope

Scandinavia

SouthernEurope

Ukraine

WesternEurope

Congo

EastAfrica

Egypt

Madagascar

NorthAfrica

SouthAfrica

Afghanistan

China

India

Irkutsk

Japan

Kamchatka

MiddleEast

Mongolia

Siam

Siberia

Ural

Yakutsk

EasternAustralia

Indonesia

NewGuinea

WesternAustralia

type Continent = [Country]

a Continent is simply a list of the Countries it contains.

neighbours :: Country -> [Country]

Given a Country, returns the list of all neighbouring countries (in no specific order).

isNeighbour :: Country -> Country -> Bool

Given two countries, returns True if there is an edge connecting them and False otherwise.

northAmerica :: Continent

A list of the countries in North America.

europe :: Continent

A list of the countries in Europe.

asia :: Continent

A list of the countries in Asia.

southAmerica :: Continent

A list of the countries in South America.

africa :: Continent

A list of the countries in Africa.

australia :: Continent

A list of the countries in Australia.