SetupBoard

Description

The is a basic representation of the Game during the initial troop placement rounds.

Documentation

data SetupState

Contains a SetupBoardState and information about how complete the board is.

Constructors

Incomplete SetupBoardState
when not all countries have a owner

PartiallyComplete SetupBoardState
when all countries have an owner but not all troops have been placed

Complete SetupBoardState
when all troops have been placed

emptyBoard :: [Player] -> SetupState

Creates a blank board with no troops and where no countries are owned yet. Creates an Incomplete SetupBoardState.

placeTroop :: Country -> SetupState -> Maybe SetupState

Partial function, only defined for incomplete and partially complete SetupState and only when the current player owns the given country. It should error if called on a Complete SetupBoardState. It should return Nothing if called on a Country not owned by current player (or unowned).

completeBoardOwner :: SetupState -> Country -> (Player, Int)

partiallyCompleteBoardOwner :: SetupState -> Country -> (Player, Int)

Partial function, defined for PartiallyComplete SetupState, gives the owner and number of troops in each country. It should error if called on a incomplete SetupState

incompleteBoardOwner :: SetupState -> Country -> (Maybe Player, Int)

Partial function, only defined for any SetupState, gives the owner and number of troops in each country.

setUpTurnOrder :: SetupState -> [Player]

Gives the turnOrder for a SetupState, with the player's whose turn it currently is at the head of the list.