State

Description

The basic representation of a Game excluding all dynamic and non-deterministic elements. This module focuses only on encapsulating all information needed to reconstruct a given game position and doesn't focus on the actual mechanics or logic of gameplay.

Documentation

data GameState

data MiniPhase

Constructors

MidBattle

Fields

Country

Country from which attack is taking place

Country

Country being attacked

Attackers

Number of attacking troops

WonBattle Country Country Attackers

TimeToTrade

Normal

data Phase

Constructors

Reinforce

Attack MiniPhase

Fortify

newGame :: [Player] -> (Country -> (Player, Int)) -> StdGen -> GameState

Creates a blank game with no troops in any country. Players should be given in turn order starting with the current player.

troops :: GameState -> Country -> Int

Gives the number of troops in a given country.

turnOrder :: GameState -> [Player]

Returns a list of players (in the order of play) starting with the current player.

owner :: GameState -> Country -> Player

Gives the player who owns a certain country.

changeTroops :: Country -> Int -> GameState -> GameState

Modifies (adds or subtracts) the given number of troops (to or from a country).

changeOwner :: Country -> Player -> GameState -> GameState

Replaces the owner of a country.

nextTurn :: GameState -> GameState

Advances to the next turn (and updates the phase).

currentStdGen :: GameState -> StdGen

Gets the current StdGen.

updateStdGen :: StdGen -> GameState -> GameState

Replaces the current StdGen.

phase :: GameState -> Phase

Gets the current Phase.

nextPhase :: GameState -> GameState

Updates the current Phase.

changeMiniPhase :: MiniPhase -> GameState -> GameState

Does nothing if not in Attack Phase, otherwise sets phase to Attack MiniPhase (inserting the provided MiniPhase). Note that calling this function on a WonBattle will cause it to remember that the current player gets a card this turn.

cards :: GameState -> Player -> [Card]

Returns the list of cards in the player's hand

useCard :: Player -> Card -> GameState -> GameState

Puts the card in the discard pile.

drawCard :: Player -> GameState -> GameState

Adds the top card to the player's hand. Shuffles the discard pile if necessary using a provided shuffle function.

kick :: Player -> Player -> GameState -> GameState

Removes the Player given in the second argument from the game. Gives all cards to the Player given in the first argument.

hasDrawn :: GameState -> Bool

True if and only if the current player has drawn a card this turn.